requirements.txt
# CORE
Django==4.2.3
whitenoise==6.8.2
gunicorn==20.1.0
django-environ==0.10.0
django-guardian==2.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.2.0
django-redis==5.3.0
redis==4.6.0
celery==5.3.1
sentry-sdk==1.28.1


# STANDARD PACKAGES
django-filter==23.2
django-nested-admin==4.0.2
drf-standardized-errors[openapi]==0.12.5

# translations
django-modeltranslation==0.19.16
googletrans==4.0.0rc1

# MODULE PACKAGES
django-anymail[sendgrid]==10.0
twilio==9.6.3
firebase-admin==6.2.0

# SWAGGER
drf-spectacular==0.26.3
drf-spectacular[sidecar]


flake8==6.0.0
Pillow==10.0.0
psycopg==3.1.9
PyJWT==2.7.0
pytz==2023.3

google-auth==2.22.0
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.2
google-cloud-vision==3.8.1
google-resumable-media==2.5.0
googleapis-common-protos==1.58.0
django-storages==1.13.2
django-stdimage==6.0.1
hashids==1.3.1

# protobuf>=3.11.2,<3.12.0
inapppy @ git+https://github.com/development-at-pixelfield/InAppPy.git

coverage==7.2.7
django-leaflet
loguru==0.7.2
pydantic==2.10.2

ximilar-client
django-resized

python-Levenshtein==0.26.1

app-store-server-library
django-unfold
pyairtable
googlemaps
requests>=2.31.0
